+++ src/CslaGeneratorSerialization.Analysis/Extensions/SerializationItemDefinition.cs
@@ -1,0 +1,7 @@
+﻿using CslaGeneratorSerialization.Analysis.Models;
+using Microsoft.CodeAnalysis;
+
+namespace CslaGeneratorSerialization.Analysis.Extensions;
+
+internal sealed record SerializationItemDefinition(
+	string PropertyInfoFieldName, INamedTypeSymbol PropertyInfoContainingType, ITypeSymbol PropertyInfoDataType);